feat: rename project
