+++ build.gradle
@@ -4,7 +4,7 @@ plugins {
 	id 'java'
 }
 
-group = 'net.murreng'
+group = 'net.mureng'
 version = '0.0.1-SNAPSHOT'
 sourceCompatibility = '11'
