+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'murreng'
+rootProject.name = 'mureng'
