+++ src/main/java/net/mureng/mureng/MurengBackendApplication.java
@@ -1,13 +1,13 @@
-package net.murreng.murreng;
+package net.mureng.mureng;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class MurrengBackendApplication {
+public class MurengBackendApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(MurrengBackendApplication.class, args);
+		SpringApplication.run(MurengBackendApplication.class, args);
 	}
 
 }
